Logout with client id
Logout with client_id parameter is allowed due to openid specification. When id_token_hint is not specified, client_id can be used instead. This change is a solution for situation when session_token is expired and in the logout operation we get the error message "Missing Parameters: id_token_hint error on Session timeout.".